Display a product's reviews via modal.
A modal dialog is created for each product listed in /products,
and a product's reviews modal is revealed upon clicking "Reviews".

This results in multiple, unnecessary and mostly un-viewed modals
rendered on the page; killing page load times.

Next we'll replace this with Ajax generation of the modal upon demand.